Next & Previous button function
After feedback on how to get the next button working, I updated it to work through the entire API. There was a break between entries 898 and 10001, so I needed to fix that bug. Additionally, I wanted something for reaching the beginning or the end of the index.

After updating the "showNextPokemon" function, I wrote the "showPrevPokemon" function to work in the opposite direction.